+++ crates/orchestrator/src/tests/workers/update_state/mod.rs
@@ -33,15 +33,18 @@ async fn test_update_state_worker(
     // Mocking db function expectations
     // If no successful state update jobs exist
     if !last_successful_job_exists {
-        db.expect_get_last_successful_job_by_type().with(eq(JobType::StateTransition)).times(1).returning(|_| Ok(None));
+        db.expect_get_latest_job_by_type_and_status()
+            .with(eq(JobType::StateTransition), eq(JobStatus::Completed))
+            .times(1)
+            .returning(|_, _| Ok(None));
     } else {
         // if successful state update job exists
 
         // mocking the return value of first function call (getting last successful jobs):
-        db.expect_get_last_successful_job_by_type()
-            .with(eq(JobType::StateTransition))
+        db.expect_get_latest_job_by_type_and_status()
+            .with(eq(JobType::StateTransition), eq(JobStatus::Completed))
             .times(1)
-            .returning(|_| Ok(Some(get_job_item_mock_by_id("1".to_string(), Uuid::new_v4()))));
+            .returning(|_, _| Ok(Some(get_job_item_mock_by_id("1".to_string(), Uuid::new_v4()))));
 
         // mocking the return values of second function call (getting completed proving worker jobs)
         db.expect_get_jobs_after_internal_id_by_job_type()